+++ opentelemetry-sdk/src/opentelemetry/sdk/metrics/_internal/__init__.py
@@ -14,11 +14,12 @@
 
 import weakref
 from atexit import register, unregister
+from dataclasses import dataclass
 from logging import getLogger
 from os import environ
 from threading import Lock
 from time import time_ns
-from typing import TYPE_CHECKING, Optional, Sequence
+from typing import TYPE_CHECKING, Callable, Optional, Sequence
 
 # This kind of import is needed to avoid Sphinx errors.
 from opentelemetry.metrics import Counter as APICounter
@@ -61,7 +62,10 @@ from opentelemetry.sdk.metrics._internal.sdk_configuration import (
     SdkConfiguration,
 )
 from opentelemetry.sdk.resources import Resource
-from opentelemetry.sdk.util.instrumentation import InstrumentationScope
+from opentelemetry.sdk.util.instrumentation import (
+    InstrumentationScope,
+    _InstrumentationScopePredicateT,
+)
 from opentelemetry.util._once import Once
 from opentelemetry.util.types import (
     Attributes,
@@ -73,6 +77,27 @@ if TYPE_CHECKING:
 _logger = getLogger(__name__)
 
 
+@dataclass
+class _MeterConfig:
+    is_enabled: bool = True
+
+    @classmethod
+    def default(cls) -> "_MeterConfig":
+        return _MeterConfig()
+
+
+class _ProxyMeterConfig:
+    def __init__(self, config: _MeterConfig):
+        self._config = config
+
+    @property
+    def is_enabled(self) -> bool:
+        return self._config.is_enabled
+
+    def update(self, config: _MeterConfig) -> None:
+        self._config = config
+
+
 class Meter(APIMeter):
     """See `opentelemetry.metrics.Meter`."""
 
@@ -80,6 +105,8 @@ class Meter(APIMeter):
         self,
         instrumentation_scope: InstrumentationScope,
         measurement_consumer: MeasurementConsumer,
+        *,
+        _meter_config: Optional[_MeterConfig] = None,
     ):
         super().__init__(
             name=instrumentation_scope.name,
@@ -90,6 +117,15 @@ class Meter(APIMeter):
         self._measurement_consumer = measurement_consumer
         self._instrument_id_instrument = {}
         self._instrument_registration_lock = Lock()
+        self._meter_config = _ProxyMeterConfig(
+            _meter_config or _MeterConfig.default()
+        )
+
+    def _is_enabled(self) -> bool:
+        return self._meter_config.is_enabled
+
+    def _set_meter_config(self, meter_config: _MeterConfig) -> None:
+        self._meter_config.update(meter_config)
 
     def create_counter(self, name, unit="", description="") -> APICounter:
         with self._instrument_registration_lock:
@@ -104,6 +140,7 @@ class Meter(APIMeter):
                         self._measurement_consumer,
                         unit,
                         description,
+                        _meter_config=self._meter_config,
                     )
                 )
             instrument = self._instrument_id_instrument[status.instrument_id]
@@ -136,6 +173,7 @@ class Meter(APIMeter):
                         self._measurement_consumer,
                         unit,
                         description,
+                        _meter_config=self._meter_config,
                     )
                 )
             instrument = self._instrument_id_instrument[status.instrument_id]
@@ -173,6 +211,7 @@ class Meter(APIMeter):
                         callbacks,
                         unit,
                         description,
+                        _meter_config=self._meter_config,
                     )
                 )
             instrument = self._instrument_id_instrument[status.instrument_id]
@@ -241,6 +280,7 @@ class Meter(APIMeter):
                         unit,
                         description,
                         explicit_bucket_boundaries_advisory,
+                        _meter_config=self._meter_config,
                     )
                 )
             instrument = self._instrument_id_instrument[status.instrument_id]
@@ -268,6 +308,7 @@ class Meter(APIMeter):
                     self._measurement_consumer,
                     unit,
                     description,
+                    _meter_config=self._meter_config,
                 )
             instrument = self._instrument_id_instrument[status.instrument_id]
 
@@ -300,6 +341,7 @@ class Meter(APIMeter):
                         callbacks,
                         unit,
                         description,
+                        _meter_config=self._meter_config,
                     )
                 )
             instrument = self._instrument_id_instrument[status.instrument_id]
@@ -338,6 +380,7 @@ class Meter(APIMeter):
                         callbacks,
                         unit,
                         description,
+                        _meter_config=self._meter_config,
                     )
                 )
             instrument = self._instrument_id_instrument[status.instrument_id]
@@ -370,6 +413,45 @@ def _get_exemplar_filter(exemplar_filter: str) -> ExemplarFilter:
         return AlwaysOffExemplarFilter()
     msg = f"Unknown exemplar filter '{exemplar_filter}'."
     raise ValueError(msg)
+
+
+_MeterConfiguratorT = Callable[[InstrumentationScope], _MeterConfig]
+_MeterConfiguratorRulesT = Sequence[
+    tuple[_InstrumentationScopePredicateT, _MeterConfig]
+]
+
+
+def _default_meter_configurator(
+    _meter_scope: InstrumentationScope,
+) -> _MeterConfig:
+    return _MeterConfig.default()
+
+
+def _disable_meter_configurator(
+    _meter_scope: InstrumentationScope,
+) -> _MeterConfig:
+    return _MeterConfig(is_enabled=False)
+
+
+class _RuleBasedMeterConfigurator:
+    def __init__(
+        self,
+        *,
+        rules: _MeterConfiguratorRulesT,
+        default_config: _MeterConfig,
+    ):
+        self._rules = rules
+        self._default_config = default_config
+
+    def __call__(self, meter_scope: InstrumentationScope) -> _MeterConfig:
+        for predicate, meter_config in self._rules:
+            if predicate(meter_scope):
+                return meter_config
+        # by default return default config
+        return self._default_config
+
+    def update_rules(self, rules: _MeterConfiguratorRulesT) -> None:
+        self._rules = rules
 
 
 class MeterProvider(APIMeterProvider):
@@ -428,6 +510,8 @@ class MeterProvider(APIMeterProvider):
         exemplar_filter: Optional[ExemplarFilter] = None,
         shutdown_on_exit: bool = True,
         views: Sequence["opentelemetry.sdk.metrics.view.View"] = (),
+        *,
+        _meter_configurator: Optional[_MeterConfiguratorT] = None,
     ):
         self._lock = Lock()
         self._meter_lock = Lock()
@@ -454,9 +538,12 @@ class MeterProvider(APIMeterProvider):
         if shutdown_on_exit:
             self._atexit_handler = register(self.shutdown)
 
-        self._meters = {}
+        self._meters: dict[InstrumentationScope, Meter] = {}
         self._shutdown_once = Once()
         self._shutdown = False
+        self._meter_configurator = (
+            _meter_configurator or _default_meter_configurator
+        )
 
         for metric_reader in self._sdk_config.metric_readers:
             with self._all_metric_readers_lock:
@@ -473,6 +560,36 @@ class MeterProvider(APIMeterProvider):
                 self._measurement_consumer.collect
             )
             metric_reader._set_meter_provider(self)
+
+    def _set_meter_configurator(
+        self, *, meter_configurator: _MeterConfiguratorT
+    ):
+        """Set a new MeterConfigurator for this MeterProvider.
+
+        Setting a new MeterConfigurator will result in the configurator being called
+        for each outstanding Meter and for any newly created meters thereafter.
+        Therefore, it is important that the provided function returns quickly.
+        """
+        with self._meter_lock:
+            self._meter_configurator = meter_configurator
+            for instrumentation_scope, meter in self._meters.items():
+                # pylint: disable-next=protected-access
+                meter._set_meter_config(
+                    self._apply_meter_configurator(instrumentation_scope)
+                )
+
+    def _apply_meter_configurator(
+        self, instrumentation_scope: InstrumentationScope
+    ) -> _MeterConfig:
+        try:
+            return self._meter_configurator(instrumentation_scope)
+        # pylint: disable-next=broad-exception-caught
+        except Exception:
+            _logger.exception(
+                "meter configurator failed for scope '%s', using default config",
+                instrumentation_scope.name,
+            )
+            return _MeterConfig.default()
 
     def force_flush(self, timeout_millis: float = 10_000) -> bool:
         deadline_ns = time_ns() + timeout_millis * 10**6
@@ -554,11 +671,9 @@ class MeterProvider(APIMeterProvider):
 
             # pylint: disable=broad-exception-raised
             raise Exception(
-                (
-                    "MeterProvider.shutdown failed because the following "
-                    "metric readers failed during shutdown:\n"
-                    f"{metric_reader_error_string}"
-                )
+                "MeterProvider.shutdown failed because the following "
+                "metric readers failed during shutdown:\n"
+                f"{metric_reader_error_string}"
             )
 
     def get_meter(
@@ -581,13 +696,18 @@ class MeterProvider(APIMeterProvider):
             _logger.warning("Meter name cannot be None or empty.")
             return NoOpMeter(name, version=version, schema_url=schema_url)
 
-        info = InstrumentationScope(name, version, schema_url, attributes)
+        instrumentation_scope = InstrumentationScope(
+            name, version, schema_url, attributes
+        )
         with self._meter_lock:
-            if not self._meters.get(info):
+            if not self._meters.get(instrumentation_scope):
                 # FIXME #2558 pass SDKConfig object to meter so that the meter
                 # has access to views.
-                self._meters[info] = Meter(
-                    info,
+                self._meters[instrumentation_scope] = Meter(
+                    instrumentation_scope,
                     self._measurement_consumer,
+                    _meter_config=self._apply_meter_configurator(
+                        instrumentation_scope
+                    ),
                 )
-            return self._meters[info]
+            return self._meters[instrumentation_scope]