+++ opentelemetry-sdk/src/opentelemetry/sdk/metrics/_internal/instrument.py
@@ -38,8 +38,12 @@ from opentelemetry.metrics._internal.instrument import (
 from opentelemetry.sdk.metrics._internal.measurement import Measurement
 
 if TYPE_CHECKING:
-    import opentelemetry.sdk.metrics
+    from opentelemetry.sdk.metrics._internal import (
+        MeasurementConsumer,
+        _ProxyMeterConfig,
+    )
     from opentelemetry.sdk.util.instrumentation import InstrumentationScope
+
 
 _logger = getLogger(__name__)
 
@@ -54,9 +58,11 @@ class _Synchronous:
         self,
         name: str,
         instrumentation_scope: InstrumentationScope,
-        measurement_consumer: "opentelemetry.sdk.metrics.MeasurementConsumer",
+        measurement_consumer: MeasurementConsumer,
         unit: str = "",
         description: str = "",
+        *,
+        _meter_config: _ProxyMeterConfig | None = None,
     ):
         # pylint: disable=no-member
         result = self._check_name_unit_description(name, unit, description)
@@ -78,7 +84,11 @@ class _Synchronous:
         self.description = description
         self.instrumentation_scope = instrumentation_scope
         self._measurement_consumer = measurement_consumer
+        self._meter_config = _meter_config
         super().__init__(name, unit=unit, description=description)
+
+    def _is_enabled(self) -> bool:
+        return self._meter_config is None or self._meter_config.is_enabled
 
 
 class _Asynchronous:
@@ -86,10 +96,12 @@ class _Asynchronous:
         self,
         name: str,
         instrumentation_scope: InstrumentationScope,
-        measurement_consumer: "opentelemetry.sdk.metrics.MeasurementConsumer",
+        measurement_consumer: MeasurementConsumer,
         callbacks: Iterable[CallbackT] | None = None,
         unit: str = "",
         description: str = "",
+        *,
+        _meter_config: _ProxyMeterConfig | None = None,
     ):
         # pylint: disable=no-member
         result = self._check_name_unit_description(name, unit, description)
@@ -111,6 +123,7 @@ class _Asynchronous:
         self.description = description
         self.instrumentation_scope = instrumentation_scope
         self._measurement_consumer = measurement_consumer
+        self._meter_config = _meter_config
         super().__init__(name, callbacks, unit=unit, description=description)
 
         self._callbacks: List[CallbackT] = []
@@ -134,9 +147,14 @@ class _Asynchronous:
                 else:
                     self._callbacks.append(callback)
 
+    def _is_enabled(self) -> bool:
+        return self._meter_config is None or self._meter_config.is_enabled
+
     def callback(
         self, callback_options: CallbackOptions
     ) -> Iterable[Measurement]:
+        if not self._is_enabled():
+            return
         for callback in self._callbacks:
             try:
                 for api_measurement in callback(callback_options):
@@ -165,6 +183,10 @@ class Counter(_Synchronous, APICounter):
         attributes: dict[str, str] | None = None,
         context: Context | None = None,
     ):
+        if not self._is_enabled():
+            super().add(amount, attributes=attributes, context=context)
+            return
+
         if amount < 0:
             _logger.warning(
                 "Add amount must be non-negative on Counter %s.", self.name
@@ -194,6 +216,10 @@ class UpDownCounter(_Synchronous, APIUpDownCounter):
         attributes: dict[str, str] | None = None,
         context: Context | None = None,
     ):
+        if not self._is_enabled():
+            super().add(amount, attributes=attributes, context=context)
+            return
+
         time_unix_nano = time_ns()
         self._measurement_consumer.consume_measurement(
             Measurement(
@@ -229,10 +255,12 @@ class Histogram(_Synchronous, APIHistogram):
         self,
         name: str,
         instrumentation_scope: InstrumentationScope,
-        measurement_consumer: "opentelemetry.sdk.metrics.MeasurementConsumer",
+        measurement_consumer: MeasurementConsumer,
         unit: str = "",
         description: str = "",
         explicit_bucket_boundaries_advisory: Sequence[float] | None = None,
+        *,
+        _meter_config: _ProxyMeterConfig | None = None,
     ):
         super().__init__(
             name,
@@ -240,6 +268,7 @@ class Histogram(_Synchronous, APIHistogram):
             description=description,
             instrumentation_scope=instrumentation_scope,
             measurement_consumer=measurement_consumer,
+            _meter_config=_meter_config,
         )
         self._advisory = _MetricsHistogramAdvisory(
             explicit_bucket_boundaries=explicit_bucket_boundaries_advisory
@@ -256,6 +285,10 @@ class Histogram(_Synchronous, APIHistogram):
         attributes: dict[str, str] | None = None,
         context: Context | None = None,
     ):
+        if not self._is_enabled():
+            super().record(amount, attributes=attributes, context=context)
+            return
+
         if amount < 0:
             _logger.warning(
                 "Record amount must be non-negative on Histogram %s.",
@@ -286,6 +319,10 @@ class Gauge(_Synchronous, APIGauge):
         attributes: dict[str, str] | None = None,
         context: Context | None = None,
     ):
+        if not self._is_enabled():
+            super().set(amount, attributes=attributes, context=context)
+            return
+
         time_unix_nano = time_ns()
         self._measurement_consumer.consume_measurement(
             Measurement(