+++ opentelemetry-sdk/src/opentelemetry/sdk/metrics/_internal/measurement_consumer.py
@@ -66,7 +66,7 @@ class SynchronousMeasurementConsumer(MeasurementConsumer):
         self._sdk_config = sdk_config
         # should never be mutated
         self._reader_storages: Mapping[
-            "opentelemetry.sdk.metrics.MetricReader", MetricReaderStorage
+            opentelemetry.sdk.metrics.export.MetricReader, MetricReaderStorage
         ] = {
             reader: MetricReaderStorage(
                 sdk_config,
@@ -76,7 +76,7 @@ class SynchronousMeasurementConsumer(MeasurementConsumer):
             for reader in sdk_config.metric_readers
         }
         self._async_instruments: List[
-            "opentelemetry.sdk.metrics._internal.instrument._Asynchronous"
+            opentelemetry.sdk.metrics._internal.instrument._Asynchronous
         ] = []
 
     def consume_measurement(self, measurement: Measurement) -> None: